requirements.txt
requests>=2.31.0
qdrant-client>=1.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
flask==2.3.3
flask-cors>=4.0.0